+++ ProjectPanda/ProjectPanda/Views/AppoinmentViews/ConfirmationPage.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="ProjectPanda.Views.PopUpPages.ConfirmationPage"
+             x:Class="ProjectPanda.Views.AppoinmentViews.ConfirmationPage"
              Title="Confirmation"
              BackgroundColor="{StaticResource BackgroundColor}"           
              >
