--- ProjectPanda/ProjectPanda/Views/PopUpPages/EditorPopUp.xaml
@@ -1,47 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<pages:PopupPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="ProjectPanda.Views.PopUpPages.EditorPopUp"
-             xmlns:pages="clr-namespace:Rg.Plugins.Popup.Pages;assembly=Rg.Plugins.Popup"
-             xmlns:animations="clr-namespace:Rg.Plugins.Popup.Animations;assembly=Rg.Plugins.Popup"
-            xmlns:CustomEntryEditor="clr-namespace:ProjectPanda.Views.CustomEntry"
-            >
-    
-
-        <pages:PopupPage.Animation>
-              PositionIn="Center"
-              PositionOut="Center"
-              ScaleIn="1.2"
-              ScaleOut="0.8"
-              DurationIn="400"
-              DurationOut="300"
-              EasingIn="SinOut"
-              EasingOut="SinIn"
-             HasBackgroundAnimation="True"/>
-    </pages:PopupPage.Animation>
-
-
-    <Grid VerticalOptions="StartAndExpand" HorizontalOptions="StartAndExpand" Padding="15"> 
-        <Grid.RowDefinitions>
-            <RowDefinition Height="250"/>
-            <RowDefinition Height="50"/>
-        </Grid.RowDefinitions>
-
-        
-        <!--#region The text editor-->
-        <Frame CornerRadius="8" HeightRequest="250" WidthRequest="300">
-            <CustomEntryEditor:EntryBlock/>
-        </Frame>
-    <!--#endregion-->
-    
-        <!--#region The confirm button-->
-
-        <Button Text="Confirm" HorizontalOptions="Center" Margin="10" Grid.Row="1"/>
-
-        <!--#endregion-->
-        
-        
-    </Grid>
-    
-  
-</pages:PopupPage>
