--- ProjectPanda/ProjectPanda/Views/PopUpPages/EditorPopUp.xaml.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Rg.Plugins.Popup.Pages;
-using Rg.Plugins.Popup.Services;
-using Xamarin.Forms;
-using Xamarin.Forms.Xaml;
-
-namespace ProjectPanda.Views.PopUpPages
-{
-	[XamlCompilation(XamlCompilationOptions.Compile)]
-	public partial class EditorPopUp : PopupPage
-	{
-		public EditorPopUp ()
-		{
-			InitializeComponent ();
-		}
-
-
-
-    }
-}
